backend/backendBO2/requirements.txt
flask
flask-restful
flask-jwt-extended
pika  
python-dotenv 
torch==2.6.0
unsloth==2025.3.3
transformers==4.32.0
xformers==0.0.29.post3
numpy==1.24.3 
llama-cpp-python
xgboost
pandas
numpy
scikit-learn